Supply/requirements.txt
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cryptography==2.8
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.7
Django==3.0.3
django-allauth==0.41.0
django-braces==1.14.0
django-channels==0.7.0
django-countries==6.0
django-crispy-forms==1.9.0
django-google-maps==0.12.1
django-icons==1.1.1
django-ip-geolocation==1.5.3
django-leaflet==0.26.0
django-map-widgets==0.2.3
django-model-utils==4.0.0
django-postgresql==0.0.3
django-ratings==0.3.7
django-restframework==0.0.1
django-star-ratings==0.9.1
entrypoints==0.3
future==0.18.2
geocoder==1.38.1
geojson==2.5.0
geomet==0.2.1.post1
gitdb==4.0.2
GitPython==3.1.0
gmaps==0.9.0
html2text==2020.1.16
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
oauthlib==3.1.0
ogr==0.11.0
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
Pillow==7.0.0
postgis==1.0.4
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psycopg2==2.8.4
pycparser==2.20
PyGithub==1.46
Pygments==2.6.1
PyJWT==1.7.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-gitlab==2.1.2
python3-openid==3.1.0
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==19.0.0
ratelim==0.1.6
requests==2.23.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
sentinelsat==0.13
six==1.14.0
smmap==3.0.1
sqlparse==0.3.1
swapper==1.1.2.post1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.1.0
